avoid int -> size_t conversion
Former-commit-id: 36a441d705ebc7f3b4b04f0423b2aa0498a88e78

+++ include/simple/client.hpp
@@ -75,7 +75,7 @@ public:
   }
 
 private:
-  bool request(uint8_t* data, const int data_size, T& msg)
+  bool request(uint8_t* data, const size_t data_size, T& msg)
   {
     bool success{false};
